requirements.txt
numpy
tensorflow==1.7
keras
pandas
matplotlib
jupyter
jupyterlab
tensorflow-serving-api